+++ CHANGELOG.md
@@ -10,6 +10,9 @@ Changelog
 - Implemented all notifications to be on the top of the app and in front of any other element ([PR 1748](https://github.com/input-output-hk/daedalus/pull/1748))
 - Improved network screen with responsive main copy box ([PR 1797](https://github.com/input-output-hk/daedalus/pull/1797))
 
+### Fixes
+- Added wallet legacy/new types to the route path, in order to handle legacy and new wallets with same IDs ([PR 1803](https://github.com/input-output-hk/daedalus/pull/1803))
+
 ### Chores
 - Updated `cardano-wallet` to revision `132a5faf8c076eb8e873857a15e80db56be7e6ca` ([PR 1740](https://github.com/input-output-hk/daedalus/pull/1740))